# apps/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "SummerGear Go API",
    "version": "0.6.0",
    "description": "Go service-session authentication, listings, and private listing image lifecycle. Social provider credentials and Storage credentials may remain disabled in isolated test environments."
  },
  "paths": {
...
    "/api/v1/me/seller-application": {
      "get": { "operationId": "mySellerStatus", "description": "Current member's latest seller application, active approved seller, and DB-backed operator role.", "responses": { "200": { "description": "Seller status.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SellerStatus" } } } }, "401": { "$ref": "#/components/responses/SessionRequired" }, "503": { "$ref": "#/components/responses/ServiceUnavailable" } } },
      "post": { "operationId": "applySeller", "description": "Member applies; approval is operator-only and a pending application cannot be duplicated.", "parameters": [{ "$ref": "#/components/parameters/CSRFToken" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["type", "displayName"], "additionalProperties": false, "properties": { "type": { "type": "string", "enum": ["individual", "business"] }, "displayName": { "type": "string", "minLength": 1, "maxLength": 120 } } } } } }, "responses": { "201": { "description": "Pending application.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SellerApplication" } } } }, "400": { "$ref": "#/components/responses/InvalidRequest" }, "401": { "$ref": "#/components/responses/SessionRequired" }, "409": { "$ref": "#/components/responses/Conflict" } } }
    },
    "/api/v1/seller-applications": { "get": { "operationId": "pendingSellerApplications", "description": "DB-backed operator queue.", "responses": { "200": { "description": "Pending applications.", "content": { "application/json": { "schema": { "type": "object", "required": ["items"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/SellerApplication" } } } } } } }, "401": { "$ref": "#/components/responses/SessionRequired" }, "403": { "$ref": "#/components/responses/Forbidden" } } } },
...
    "/api/v1/listings/{id}/inventory": {
      "get": { "operationId": "ownerInventory", "description": "Listing owner sees registered stock or null.", "parameters": [{ "$ref": "#/components/parameters/ListingID" }], "responses": { "200": { "description": "Inventory.", "content": { "application/json": { "schema": { "type": "object", "required": ["inventory"], "properties": { "inventory": { "oneOf": [{ "type": "null" }, { "$ref": "#/components/schemas/InventoryView" }] } } } } } }, "404": { "$ref": "#/components/responses/NotFound" } } },
      "put": { "operationId": "setOwnerInventory", "description": "Approved seller and active owner listing only. Unit-one stock is not replenished while reserved or consumed.", "parameters": [{ "$ref": "#/components/parameters/ListingID" }, { "$ref": "#/components/parameters/CSRFToken" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["availableQuantity"], "additionalProperties": false, "properties": { "availableQuantity": { "type": "integer", "enum": [0, 1] } } } } } }, "responses": { "200": { "description": "Updated inventory.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InventoryView" } } } }, "400": { "$ref": "#/components/responses/InvalidRequest" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" } } }
    },
...
    "/api/v1/me/push-devices": {
      "post": { "operationId": "registerGoPushDevice", "description": "Register a native Expo token against the current Go auth session. Re-registering transfers this physical token to the current member. Logout/expiry prevents delivery.", "parameters": [{ "$ref": "#/components/parameters/CSRFToken" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PushDeviceRequest" } } } }, "responses": { "200": { "description": "Device registered.", "content": { "application/json": { "schema": { "type": "object", "required": ["registered"], "properties": { "registered": { "const": true } } } } } }, "400": { "$ref": "#/components/responses/InvalidRequest" }, "401": { "$ref": "#/components/responses/SessionRequired" }, "403": { "$ref": "#/components/responses/Forbidden" }, "503": { "$ref": "#/components/responses/ServiceUnavailable" } } },
      "delete": { "operationId": "unregisterGoPushDevice", "description": "Disable this device for the current Go session; CSRF and origin required.", "parameters": [{ "$ref": "#/components/parameters/CSRFToken" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PushDeviceRequest" } } } }, "responses": { "204": { "description": "Device disabled or already absent." }, "400": { "$ref": "#/components/responses/InvalidRequest" }, "401": { "$ref": "#/components/responses/SessionRequired" }, "403": { "$ref": "#/components/responses/Forbidden" } } }
    },
...
  },
  "components": {
    "responses": {
      "InvalidRequest": { "description": "Invalid request.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "SessionRequired": { "description": "Session required.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Forbidden": { "description": "Reviewer access or origin/CSRF denied.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "Listing unavailable to requester.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Conflict": { "description": "State conflict.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "ServiceUnavailable": { "description": "Backing database or dependency unavailable.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    },
    "parameters": {
...
      "ListingID": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "CSRFToken": {
        "name": "X-CSRF-Token",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "schemas": {
      "SellerApplication": { "type": "object", "additionalProperties": false, "required": ["id", "applicantId", "type", "displayName", "status", "sellerId", "reason", "reviewedBy", "createdAt", "reviewedAt"], "properties": { "id": { "type": "string", "format": "uuid" }, "applicantId": { "type": "string", "format": "uuid" }, "type": { "type": "string", "enum": ["individual", "business"] }, "displayName": { "type": "string" }, "status": { "type": "string", "enum": ["pending", "approved", "rejected"] }, "sellerId": { "type": ["string", "null"], "format": "uuid" }, "reason": { "type": ["string", "null"] }, "reviewedBy": { "type": ["string", "null"], "format": "uuid" }, "createdAt": { "type": "string", "format": "date-time" }, "reviewedAt": { "type": ["string", "null"], "format": "date-time" } } },
      "SellerStatus": { "type": "object", "additionalProperties": false, "required": ["application", "seller", "reviewer"], "properties": { "application": { "oneOf": [{ "type": "null" }, { "$ref": "#/components/schemas/SellerApplication" }] }, "seller": { "oneOf": [{ "type": "null" }, { "type": "object", "required": ["id", "displayName", "status"], "properties": { "id": { "type": "string", "format": "uuid" }, "displayName": { "type": "string" }, "status": { "type": "string", "enum": ["approved"] } } }] }, "reviewer": { "type": "boolean" } } },
      "InventoryView": { "type": "object", "additionalProperties": false, "required": ["listingId", "sellerId", "availableQuantity", "reservedQuantity"], "properties": { "listingId": { "type": "string", "format": "uuid" }, "sellerId": { "type": "string", "format": "uuid" }, "availableQuantity": { "type": "integer", "minimum": 0, "maximum": 1 }, "reservedQuantity": { "type": "integer", "minimum": 0 } } },
...
      "PushDeviceRequest": { "type": "object", "additionalProperties": false, "required": ["token", "platform"], "properties": { "token": { "type": "string", "pattern": "^(Expo|Exponent)PushToken\\[[a-zA-Z0-9_-]{10,240}\\]$" }, "platform": { "type": "string", "enum": ["ios", "android"] } } },
...
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "pattern": "^[a-f0-9]{32}$"
          }
        }
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-sg_session"
      }
    }
  }
}